third_party/WebKit/LayoutTests/fast/dom/shadow/drop-event-in-shadow-expected.txt

CONSOLE WARNING: line 24: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS successfullyParsed is true

TEST COMPLETE

PASS: drop event is fired.
